enabling for both vim/nvim

+++ appconfig/tmux/install.sh
@@ -4,6 +4,12 @@
 APP_PATH=`dirname "$0"`
 APP_PATH=`( cd "$APP_PATH" && pwd )`
 
+var1="18.04"
+var2=`lsb_release -r | awk '{ print $2 }'`
+if [ "$var2" = "$var1" ]; then
+  export BEAVER=1
+fi
+
 default=y
 while true; do
   [[ -t 0 ]] && { read -t 10 -n 2 -p $'\e[1;32mInstall TMUX? [y/n] (default: '"$default"$')\e[0m\n' resp || resp=$default ; }
@@ -14,28 +20,36 @@ while true; do
 
     toilet Installing tmux
 
-    libevent=`ldconfig -p | grep libevent-2.0 | wc -l`
-    if [ "$libevent" -lt "2" ]; then
+    if [ -n "$BEAVER" ]; then
 
-      echo $'\033[31mLibevent not installed, installing... \033[00m' 
+      sudo apt -y install tmux
 
-      # install libevent
-      cd /tmp
-      wget https://github.com/libevent/libevent/releases/download/release-2.0.22-stable/libevent-2.0.22-stable.tar.gz
-      tar -xvf libevent-2.0.22-stable.tar.gz
-      rm libevent-2.0.22-stable.tar.gz
-      cd libevent-2.0.22-stable
-      ./configure && make >> /dev/null
-      sudo make install
+    else
+
+      libevent=`ldconfig -p | grep libevent-2.0 | wc -l`
+      if [ "$libevent" -lt "2" ]; then
+
+        echo $'\033[31mLibevent not installed, installing... \033[00m' 
+
+        # install libevent
+        cd /tmp
+        wget https://github.com/libevent/libevent/releases/download/release-2.0.22-stable/libevent-2.0.22-stable.tar.gz
+        tar -xvf libevent-2.0.22-stable.tar.gz
+        rm libevent-2.0.22-stable.tar.gz
+        cd libevent-2.0.22-stable
+        ./configure && make >> /dev/null
+        sudo make install
+
+      fi
+
+      # instal tmux
+      cd $APP_PATH/../../submodules/tmux
+      sh autogen.sh
+      ./configure && make -j4
+      sudo make install-binPROGRAMS
+      git clean -fd
 
     fi
-
-    # instal tmux
-    cd $APP_PATH/../../submodules/tmux
-    sh autogen.sh
-    ./configure && make -j4
-    sudo make install-binPROGRAMS
-    git clean -fd
 
     break
   elif [[ $response =~ ^(n|N)=$ ]]
